Rearange file structure  - move hwloc/not-hwloc implementation to separate file
