+++ MANIFEST
@@ -47,14 +47,15 @@ include/memkind.h
 include/memkind/internal/heap_manager.h
 include/memkind/internal/memkind_arena.h
 include/memkind/internal/memkind_bitmask.h
+include/memkind/internal/memkind_capacity.h
 include/memkind/internal/memkind_dax_kmem.h
 include/memkind/internal/memkind_default.h
 include/memkind/internal/memkind_gbtlb.h
 include/memkind/internal/memkind_hbw.h
-include/memkind/internal/memkind_hmat.h
 include/memkind/internal/memkind_hugetlb.h
 include/memkind/internal/memkind_interleave.h
 include/memkind/internal/memkind_log.h
+include/memkind/internal/memkind_mem_attributes.h
 include/memkind/internal/memkind_pmem.h
 include/memkind/internal/memkind_private.h
 include/memkind/internal/memkind_regular.h
@@ -412,14 +413,15 @@ src/memkind-hbw-nodes.c
 src/memkind.c
 src/memkind_arena.c
 src/memkind_bitmask.c
+src/memkind_capacity.c
 src/memkind_dax_kmem.c
 src/memkind_default.c
 src/memkind_gbtlb.c
 src/memkind_hbw.c
-src/memkind_hmat.c
 src/memkind_hugetlb.c
 src/memkind_interleave.c
 src/memkind_log.c
+src/memkind_mem_attributes.c
 src/memkind_pmem.c
 src/memkind_regular.c
 src/tbb_wrapper.c
