+++ include/memkind/internal/memkind_capacity.h
