+++ src/memkind.c
@@ -9,8 +9,8 @@
 #include <memkind/internal/memkind_default.h>
 #include <memkind/internal/memkind_hugetlb.h>
 #include <memkind/internal/memkind_arena.h>
+#include <memkind/internal/memkind_capacity.h>
 #include <memkind/internal/memkind_hbw.h>
-#include <memkind/internal/memkind_hmat.h>
 #include <memkind/internal/memkind_regular.h>
 #include <memkind/internal/memkind_gbtlb.h>
 #include <memkind/internal/memkind_dax_kmem.h>
