+++ src/memkind_capacity.c
